Check active governor and flush STDOUT pipes
1. In function set_governor() add check for current active governor, and perform governor change to "powersave" only if it is not in "powersave" mode. Also print message to log about this.

Motivation: it saves few system calls every 2 seconds, and possibly avoids some unknown problems with epp driver(it can trigger some internal changes even if we changing to the same value?). Print message because by default governor should be in "powersave" mode, and if something changes it constantly we should see this in the logs.

2. Add sys.stdout.flush() call after every print message.
Our service's STDOUT and STDERR are pipes, and in this case the buffer is only flushed once it is full. Hence the script's messages only turn up in systemd's logs after it has produced even more output.
Alternative is to add "Environment=PYTHONUNBUFFERED=1" to service unit file in [Service] section.